--- a/benchmarks/BaroquenMelody.Benchmarks/Program.cs
+++ b/benchmarks/BaroquenMelody.Benchmarks/Program.cs
@@ -1,5 +1,4 @@
 ﻿using BaroquenMelody.Benchmarks.Compositions.Rules;
-using BenchmarkDotNet.Configs;
 using BenchmarkDotNet.Running;
 
-BenchmarkRunner.Run<AggregateCompositionRuleBenchmarks>(new DebugInProcessConfig());
+BenchmarkRunner.Run<FollowsStandardProgressionBenchmarks>();