… benchmarks/BaroquenMelody.Benchmarks/Compositions/Extensions/BaroquenChordExtensionsBenchmarks.cs
… benchmarks/BaroquenMelody.Benchmarks/Compositions/Extensions/BaroquenNoteExtensionsBenchmarks.cs
… benchmarks/BaroquenMelody.Benchmarks/Compositions/Rules/AggregateCompositionRuleBenchmarks.cs
… benchmarks/BaroquenMelody.Benchmarks/Compositions/Rules/AvoidDirectIntervalsBenchmarks.cs
… benchmarks/BaroquenMelody.Benchmarks/Compositions/Rules/AvoidDissonanceBenchmarks.cs
… benchmarks/BaroquenMelody.Benchmarks/Compositions/Rules/AvoidParallelIntervalsBenchmarks.cs
… benchmarks/BaroquenMelody.Benchmarks/Compositions/Rules/AvoidRepetitionBenchmarks.cs
… benchmarks/BaroquenMelody.Benchmarks/Compositions/Rules/FollowsStandardProgressionBenchmarks.cs
… benchmarks/BaroquenMelody.Benchmarks/Compositions/Strategies/CompositionStrategyBenchmarks.cs
… src/BaroquenMelody.Library/Compositions/Rules/CompositionRuleFactory.cs
… src/BaroquenMelody.Library/Infrastructure/Collections/Extensions/EnumerableExtensions.cs
… tests/BaroquenMelody.Library.Tests/Compositions/Composers/ChordComposerTests.cs
… tests/BaroquenMelody.Library.Tests/Compositions/Composers/ComposerTests.cs
… tests/BaroquenMelody.Library.Tests/Compositions/Rules/AvoidDirectIntervalsTest.cs
… tests/BaroquenMelody.Library.Tests/Compositions/Rules/FollowsStandardProgressionTests.cs
… tests/BaroquenMelody.Library.Tests/Compositions/Strategies/CompositionStrategyTests.cs
… tests/BaroquenMelody.Library.Tests/Infrastructure/Collections/Extensions/EnumerableExtensionsTests.cs