## Description

Various performance improvements within the hot paths of the app, using benchmarks to validate.

```
BenchmarkDotNet v0.14.0, Windows 10 (10.0.19045.4894/22H2/2022Update)
AMD Ryzen 9 3900X, 1 CPU, 24 logical and 12 physical cores
.NET SDK 8.0.400
  [Host]     : .NET 8.0.8 (8.0.824.36612), X64 RyuJIT AVX2
  DefaultJob : .NET 8.0.8 (8.0.824.36612), X64 RyuJIT AVX2
```

| Method                    | Mean     | Error   | StdDev  | Ratio | RatioSD | Gen0   | Allocated | Alloc Ratio |
|-------------------------- |---------:|--------:|--------:|------:|--------:|-------:|----------:|------------:|
| ApplyChordChoiceOptimized | 370.1 ns | 6.94 ns | 7.12 ns |  0.92 |    0.02 | 0.1183 |     992 B |        0.92 |
| ApplyChordChoice          | 401.9 ns | 7.78 ns | 7.28 ns |  1.00 |    0.03 | 0.1287 |    1080 B |        1.00 |

| Method                   | Mean     | Error    | StdDev   | Ratio | RatioSD | Gen0   | Allocated | Alloc Ratio |
|------------------------- |---------:|---------:|---------:|------:|--------:|-------:|----------:|------------:|
| ApplyNoteChoiceOptimized | 20.88 ns | 0.305 ns | 0.286 ns |  0.85 |    0.02 | 0.0124 |     104 B |        1.30 |
| ApplyNoteChoice          | 24.61 ns | 0.440 ns | 0.390 ns |  1.00 |    0.02 | 0.0095 |      80 B |        1.00 |

| Method                                | Mean     | Error    | StdDev   | Ratio | RatioSD | Gen0   | Allocated | Alloc Ratio |
|-------------------------------------- |---------:|---------:|---------:|------:|--------:|-------:|----------:|------------:|
| AvoidDirectIntervalsEvaluateOptimized | 803.6 ns |  5.84 ns |  5.18 ns |  0.87 |    0.01 |      - |         - |        0.00 |
| AvoidDirectIntervalsEvaluate          | 920.8 ns | 10.83 ns | 10.13 ns |  1.00 |    0.02 | 0.0525 |     440 B |        1.00 |

| Method                           | Mean      | Error    | StdDev   | Ratio | RatioSD | Gen0   | Allocated | Alloc Ratio |
|--------------------------------- |----------:|---------:|---------:|------:|--------:|-------:|----------:|------------:|
| AvoidDissonanceEvaluateOptimized |  64.47 ns | 0.490 ns | 0.458 ns |  0.33 |    0.00 |      - |         - |        0.00 |
| AvoidDissonanceEvaluate          | 194.21 ns | 2.959 ns | 2.623 ns |  1.00 |    0.02 | 0.0544 |     456 B |        1.00 |

| Method                           | Mean     | Error    | StdDev   | Ratio | RatioSD | Gen0   | Allocated | Alloc Ratio |
|--------------------------------- |---------:|---------:|---------:|------:|--------:|-------:|----------:|------------:|
| AvoidRepetitionEvaluateOptimized | 49.38 ns | 0.972 ns | 0.954 ns |  0.58 |    0.01 | 0.0114 |      96 B |        0.52 |
| AvoidRepetitionEvaluate          | 85.63 ns | 1.077 ns | 0.954 ns |  1.00 |    0.02 | 0.0219 |     184 B |        1.00 |

| Method                                      | Mean     | Error   | StdDev  | Ratio | Gen0   | Allocated | Alloc Ratio |
|-------------------------------------------- |---------:|--------:|--------:|------:|-------:|----------:|------------:|
| FollowsStandardProgressionEvaluateOptimized | 340.1 ns | 3.52 ns | 3.29 ns |  0.72 | 0.1335 |    1120 B |        2.33 |
| FollowsStandardProgressionEvaluate          | 472.1 ns | 4.13 ns | 3.87 ns |  1.00 | 0.0572 |     480 B |        1.00 |

| Method                                                    | Mean      | Error     | StdDev    | Ratio | RatioSD | Gen0      | Gen1    | Allocated | Alloc Ratio |
|---------------------------------------------------------- |----------:|----------:|----------:|------:|--------:|----------:|--------:|----------:|------------:|
| CompositionStrategyGetValidChordChoicesAndChordsOptimized |  9.944 ms | 0.0762 ms | 0.0636 ms |  0.19 |    0.00 | 5796.8750 | 15.6250 |  46.11 MB |        1.00 |
| CompositionStrategyGetValidChordChoicesAndChords          | 51.454 ms | 0.6772 ms | 0.6003 ms |  1.00 |    0.02 | 5700.0000 |       - |  45.95 MB |        1.00 |


## Type of Change

- [ ] Bug Fix
- [ ] New Feature
- [ ] Breaking Change
- [x] Refactor
- [ ] Documentation
- [ ] Other (please describe)

## Checklist

- [x] I have read the [contributing guidelines](https://github.com/wbaldoumas/baroquen-melody/blob/main/CONTRIBUTING.md)
- [x] Existing issues have been referenced (where applicable)
- [x] I have verified this change is not present in other open pull requests
- [x] Functionality is documented
- [x] All code style checks pass
- [x] New code contribution is covered by automated tests
- [x] All new and existing tests pass